Add golangci-lint v2 config and fix findings
The new v2 version of `golangci-lint` comes with configuration file
structure and therefore the configuration had to be migrated.
I simply had to execute `golangci-lint migrate` and copy the comments to
the new config as they are not touched by the migration. There were also
few minor new findings that had to be fixed.